worker class added to the project
this class implements a thread that can be instructed about how to
interface with other components via callbacks to ease code reusability
and independence.

Here the read- and write-out operations are performed. The Connection
class implements the I/O procedures with the serial port of the Arduino in
a (fair) mixture of synchronous and asynchronous operations implemented via
asio.

Errors and data are reported via tha callback function.